request body queue byte reading fix

+++ src/main/java/common/queue/BodyBytesDequeue.java
@@ -1,5 +1,5 @@
 package common.queue;
 
 public interface BodyBytesDequeue {
-    byte dequeue();
+    int dequeue();
 }
