+++ src/main/java/common/queue/BodyBytesEnqueue.java
@@ -1,5 +1,5 @@
 package common.queue;
 
 public interface BodyBytesEnqueue {
-    void enqueue(byte requestBodyByte);
+    void enqueue(int requestBodyByte);
 }
